Consolidate on functional options
This changes the signature of NewClient to only accept variadic Options,
and to return an error if one occurs while setting up the client.

This leaves us with the internal implementation details of the client
and its options fully encapsulated by the package, which should give us
substantial freedom to extend the API in future while minimising
breakage for users.